Refactored all drifter tests and found a small bug
